requirements.txt
babel==2.17.0
blinker==1.9.0
click==8.3.0
Flask==3.1.2
flask-babel==4.0.0
flask-cors==6.0.1
Flask-MySQLdb==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysqlclient==2.2.7
python-dotenv==1.2.1
pytz==2025.2
Werkzeug==3.1.3
gunicorn==23.0.0
